+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Renamed `runProc` -> `runConstrained`.
 - Removed `ProcessOutput.asBytes`.
 - `ToolException` has reference to the entire `PanaProcessResult`, instead of just the `stderr`.
+- `ToolEnvironment` does not expose `environment`.
 
 ## 0.21.45